0.0.2 alpha3 change rest useCase naming. change naming for recyclerView classes
