+++ build.gradle
@@ -55,8 +55,8 @@ ext {
     moduleName = 'svc'
     desc = 'Easy and intuitive pattern for Android'
 
-    kotlin_version = '1.2.41'
-    libVersionName = "0.0.2-alpha2"
+    kotlin_version = '1.2.50'
+    libVersionName = "0.0.2-alpha3"
     libVersionCode = 1
     MIN_SDK_VERSION = 16
