+++ svc/src/main/kotlin/com/naver/android/svc/recyclerview/ActionHolder.kt
@@ -16,22 +16,11 @@
 
 package com.naver.android.svc.recyclerview
 
-import android.content.Context
-import android.support.v7.widget.RecyclerView
-import android.view.LayoutInflater
 import android.view.ViewGroup
-import com.naver.android.svc.core.common.ResourceProvider
 
 /**
- * @author bs.nam@navercorp.com 2017. 6. 19..
+ * holder with user interaction(click or swipe etc.)
+ * @author bs.nam@navercorp.com 2017. 8. 16..
  */
 
-abstract class BaseSvcHolder<in T>(layoutId: Int, parent: ViewGroup)
-    : RecyclerView.ViewHolder(LayoutInflater.from(parent.context).inflate(layoutId, parent, false))
-        , ResourceProvider {
-
-    override val context: Context?
-        get() = itemView.context
-
-    abstract fun bindView(item: T, position: Int)
-}
+abstract class ActionHolder<in T, out Listener>(layoutId: Int, parent: ViewGroup, val listener: Listener) : BaseHolder<T>(layoutId, parent)
