+++ svc/src/main/kotlin/com/naver/android/svc/recyclerview/BaseHolder.kt
@@ -16,19 +16,22 @@
 
 package com.naver.android.svc.recyclerview
 
+import android.content.Context
 import android.support.v7.widget.RecyclerView
+import android.view.LayoutInflater
+import android.view.ViewGroup
+import com.naver.android.svc.core.common.ResourceProvider
 
 /**
- * @author bs.nam@navercorp.com 2017. 8. 16..
+ * @author bs.nam@navercorp.com 2017. 6. 19..
  */
 
-abstract class SvcAdapter<VH : RecyclerView.ViewHolder, out Listener>(val listener: Listener) : RecyclerView.Adapter<VH>() {
-    final override fun onBindViewHolder(holder: VH, position: Int) {
-        if (holder.adapterPosition == RecyclerView.NO_POSITION) {
-            return
-        }
-        onBindUseCaseHolder(holder, holder.adapterPosition)
-    }
+abstract class BaseHolder<in T>(layoutId: Int, parent: ViewGroup)
+    : RecyclerView.ViewHolder(LayoutInflater.from(parent.context).inflate(layoutId, parent, false))
+        , ResourceProvider {
 
-    abstract fun onBindUseCaseHolder(holder: VH, adapterPosition: Int)
+    override val context: Context?
+        get() = itemView.context
+
+    abstract fun bindView(item: T, position: Int)
 }
